Improve handling of stop and pause events.
Resuming after a pause now re-initialize all the driver owners in a way which should be thread safe.
Wait and parallel decorator and InstrTask are now more careful when dealing with a dictionary extracted from the database. Use shallow copy to prevent modification of the dict in a non-thread safe way, this might still lead to race-conditions should do better.